refactor: rename db-connection.js to database-helper.js
